+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Future release
 
+* bugfix: vacuum udf could get into a look and overwrite stripes ([#92][])
 * add columnar decompressed chunk cache ([#86][])
 * bugfix: vacuum hanging indefinitely in some cases ([#80][])
 * bugfix: release memory during long sequential scans ([#78][])
@@ -58,6 +59,7 @@
 [#78]: https://github.com/hydradatabase/hydra/pull/78
 [#80]: https://github.com/hydradatabase/hydra/pull/80
 [#86]: https://github.com/hydradatabase/hydra/pull/86
+[#92]: https://github.com/hydradatabase/hydra/pull/92
 [02d2253]: https://github.com/hydradatabase/hydra/commit/02d2253
 [0d41837]: https://github.com/hydradatabase/hydra/commit/0d41837
 [15193be]: https://github.com/hydradatabase/hydra/commit/15193be